move validate from Library to RateCollection (#433)
this is a better fit for this functionality
